[PDI-10610] - Refactoring of the IMetaverseReader implementations to a single class, BlueprintsGraphMetaverseReader.
It is accepts a Graph as a constructor arg rather than having different impls for each possible type.
Also configured the integration test to use the new reader to export the graphml file.
