+++ src/main/java/com/pentaho/metaverse/impl/MetaverseObjectFactory.java
@@ -1,0 +1,54 @@
+/*
+ * PENTAHO CORPORATION PROPRIETARY AND CONFIDENTIAL
+ *
+ * Copyright 2002 - 2014 Pentaho Corporation (Pentaho). All rights reserved.
+ *
+ * NOTICE: All information including source code contained herein is, and
+ * remains the sole property of Pentaho and its licensors. The intellectual
+ * and technical concepts contained herein are proprietary and confidential
+ * to, and are trade secrets of Pentaho and may be covered by U.S. and foreign
+ * patents, or patents in process, and are protected by trade secret and
+ * copyright laws. The receipt or possession of this source code and/or related
+ * information does not convey or imply any rights to reproduce, disclose or
+ * distribute its contents, or to manufacture, use, or sell anything that it
+ * may describe, in whole or in part. Any reproduction, modification, distribution,
+ * or public display of this information without the express written authorization
+ * from Pentaho is strictly prohibited and in violation of applicable laws and
+ * international treaties. Access to the source code contained herein is strictly
+ * prohibited to anyone except those individuals and entities who have executed
+ * confidentiality and non-disclosure agreements or other agreements with Pentaho,
+ * explicitly covering such access.
+ */
+
+package com.pentaho.metaverse.impl;
+
+import com.pentaho.dictionary.DictionaryConst;
+import com.pentaho.dictionary.MetaverseLink;
+import com.pentaho.dictionary.MetaverseTransientNode;
+import org.pentaho.platform.api.metaverse.IMetaverseDocument;
+import org.pentaho.platform.api.metaverse.IMetaverseLink;
+import org.pentaho.platform.api.metaverse.IMetaverseNode;
+import org.pentaho.platform.api.metaverse.IMetaverseObjectFactory;
+
+public class MetaverseObjectFactory implements IMetaverseObjectFactory {
+
+
+  @Override
+  public IMetaverseDocument createDocumentObject() {
+    return new MetaverseDocument( );
+  }
+
+  @Override
+  public IMetaverseLink createLinkObject() {
+    return new MetaverseLink( );
+  }
+
+  @Override
+  public IMetaverseNode createNodeObject( String id ) {
+    MetaverseTransientNode node = new MetaverseTransientNode();
+    node.setStringID( id );
+    node.setProperty( DictionaryConst.NODE_VIRTUAL, true );
+    return node;
+  }
+
+}
