+++ src/test/java/com/pentaho/metaverse/graph/BlueprintsGraphMetaverseReaderTest.java
@@ -1,46 +1,167 @@
 package com.pentaho.metaverse.graph;
 
-import java.io.File;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.OutputStream;
+import static org.junit.Assert.*;
 
-import org.apache.commons.logging.Log;
-import org.apache.commons.logging.LogFactory;
+import java.io.FileOutputStream;
+import java.util.ArrayList;
+import java.util.Iterator;
+import java.util.List;
 
 import com.pentaho.dictionary.DictionaryConst;
+import com.tinkerpop.blueprints.impls.tg.TinkerGraph;
+import org.junit.Before;
+import org.junit.Test;
+import org.pentaho.platform.api.metaverse.IMetaverseLink;
+import org.pentaho.platform.api.metaverse.IMetaverseNode;
+
+import com.pentaho.metaverse.api.IMetaverseReader;
+import com.tinkerpop.blueprints.Direction;
+import com.tinkerpop.blueprints.Edge;
 import com.tinkerpop.blueprints.Graph;
 import com.tinkerpop.blueprints.Vertex;
-import com.tinkerpop.blueprints.impls.tg.TinkerGraph;
 
-public class TestTinkerGraphMetaverseReader extends BaseGraphMetaverseReader {
+public class BlueprintsGraphMetaverseReaderTest {
 
-  private static final long serialVersionUID = -68131573133600167L;
+  private Graph graph;
 
-  public static String filePath;
-
-  private static final Log logger = LogFactory.getLog( TestTinkerGraphMetaverseReader.class );
-
-  private TinkerGraph graph;
-
-  public TestTinkerGraphMetaverseReader() throws Exception {
-    init();
+  @Before
+  public void init() {
+    graph = new TinkerGraph();
+    loadGraph( graph );
   }
 
-  @Override
-  protected Graph getGraph() {
-    return graph;
+  @Test
+  public void testGraphMetaverseReader() throws Exception {
+
+    BlueprintsGraphMetaverseReader metaverseReader = new BlueprintsGraphMetaverseReader( graph );
+
+    Graph graph = metaverseReader.getMetaverse();
+    assertNotNull( "Graph is null", graph );
+
+    GraphMLWriter writer = new GraphMLWriter();
+    writer.outputGraph( graph, new FileOutputStream( "testGraphMetaverseReader.graphml" ) );
+
+    String export = metaverseReader.export();
+    assertNotNull( "Export is null", export );
+
+    assertTrue( "Export content is wrong", export.indexOf( "data.txt" ) != -1 );
+    assertTrue( "Export content is wrong", export.indexOf( "IP Addr" ) != -1 );
+    assertTrue( "Export content is wrong", export.indexOf( "City" ) != -1 );
+    assertTrue( "Export content is wrong", export.indexOf( "Sales" ) != -1 );
+    assertTrue( "Export content is wrong", export.indexOf( "trans1.ktr" ) != -1 );
+
+    assertEquals( "Vertex count is wrong", 30, countVertices( graph ) );
+    assertEquals( "Edge count is wrong", 42, countEdges( graph ) );
   }
 
-  @Override
-  public Log getLogger() {
-    return logger;
+  @Test
+  public void testFindNode() throws Exception {
+
+    IMetaverseReader metaverseReader = new BlueprintsGraphMetaverseReader( graph );
+
+    IMetaverseNode node = metaverseReader.findNode( "data.txt" );
+
+    assertNotNull( "Node is null", node );
+    assertEquals( "Id is wrong", "data.txt", node.getStringID() );
+    assertEquals( "Type is wrong", "file", node.getType() );
+    assertEquals( "Name is wrong", "Text file: data.txt", node.getName() );
+
+    node = metaverseReader.findNode( "bogus" );
+
+    assertNull( "Node is not null", node );
+
   }
 
-  private void init() throws Exception {
+  @Test
+  public void testGetGraph() throws Exception {
+    BlueprintsGraphMetaverseReader metaverseReader = new BlueprintsGraphMetaverseReader( graph );
+    Graph g = metaverseReader.getGraph( "datasource1.table1.field1" );
+    assertNotNull( "Graph is null", g );
 
-    TinkerGraphMetaverseReader reader = new TinkerGraphMetaverseReader();
-    graph = (TinkerGraph) reader.getGraph();
+    GraphMLWriter writer = new GraphMLWriter();
+    writer.outputGraph( graph, new FileOutputStream( "testGetGraph.graphml" ) );
+
+    g = metaverseReader.getGraph( "bogus" );
+    assertNull( "Node is not null", g );
+
+  }
+
+  @Test
+  public void testFindLink() throws Exception {
+    BlueprintsGraphMetaverseReader metaverseReader = new BlueprintsGraphMetaverseReader( graph );
+
+    IMetaverseLink link = metaverseReader.findLink( "datasource1.table1.field1", "populates", "trans2.ktr;field1", Direction.OUT );
+    assertNotNull( "Link is null", link );
+    assertEquals( "Label is wrong", "populates", link.getLabel() );
+    assertEquals( "Id is wrong", "datasource1.table1.field1", link.getFromNode().getStringID() );
+    assertEquals( "Id is wrong", "trans2.ktr;field1", link.getToNode().getStringID() );
+
+    link = metaverseReader.findLink( "bogus", "populates", "trans2.ktr;field1", Direction.OUT );
+    assertNull( "Link is not null", link );
+
+    link = metaverseReader.findLink( "datasource1.table1.field1", "bogus", "trans2.ktr;field1", Direction.OUT );
+    assertNull( "Link is not null", link );
+
+    link = metaverseReader.findLink( "datasource1.table1.field1", "populates", "bogus", Direction.OUT );
+    assertNull( "Link is not null", link );
+
+    link = metaverseReader.findLink( "datasource1.table1.field1", "populates", "trans2.ktr;field1", Direction.IN );
+    assertNull( "Link is not null", link );
+
+    link = metaverseReader.findLink( "job2.kjb", "populates", "trans2.ktr;field1", Direction.OUT );
+    assertNull( "Link is not null", link );
+
+    link = metaverseReader.findLink( "trans2.ktr;field1", "populates", "datasource1.table1.field1", Direction.IN );
+    assertNotNull( "Link is null", link );
+    assertEquals( "Label is wrong", "populates", link.getLabel() );
+    assertEquals( "Id is wrong", "datasource1.table1.field1", link.getFromNode().getStringID() );
+    assertEquals( "Id is wrong", "trans2.ktr;field1", link.getToNode().getStringID() );
+
+  }
+
+  @Test
+  public void testSearch() throws Exception {
+    BlueprintsGraphMetaverseReader metaverseReader = new BlueprintsGraphMetaverseReader( graph );
+
+    List<String> types = new ArrayList<String>();
+    types.add( "trans" );
+    List<String> ids = new ArrayList<String>();
+    ids.add( "datasource1.table1.field1" );
+    Graph graph = metaverseReader.search( types, ids );
+    assertNotNull( "Node is null", graph );
+
+    GraphMLWriter writer = new GraphMLWriter();
+    writer.outputGraph( graph, new FileOutputStream( "testSearch.graphml" ) );
+
+    assertEquals( "Vertex count is wrong", 8, countVertices( graph ) );
+    assertEquals( "Edge count is wrong", 7, countEdges( graph ) );
+
+  }
+
+  private int countEdges( Graph graph ) {
+    Iterator<Edge> edges = graph.getEdges().iterator();
+    int edgeCount = 0;
+    while ( edges.hasNext() ) {
+      edgeCount++;
+      edges.next();
+    }
+    return edgeCount;
+  }
+
+  private int countVertices( Graph graph ) {
+    Iterator<Vertex> vertices = graph.getVertices().iterator();
+    int vertexCount = 0;
+    while ( vertices.hasNext() ) {
+      vertexCount++;
+      vertices.next();
+    }
+    return vertexCount;
+  }
+
+
+  // ------- Helper methods -------
+
+  private void loadGraph( Graph graph ) {
     // these are the nodes
     Vertex textFile = createVertex( "data.txt", "file", "Text file: data.txt" );
     Vertex textField1 = createVertex( "data.txt;field1", "filefield", "Text field: IP Addr" );
@@ -117,25 +238,6 @@ public class TestTinkerGraphMetaverseReader extends BaseGraphMetaverseReader {
     graph.addEdge( null, trans2Field2, table2field2, DictionaryConst.LINK_POPULATES );
     graph.addEdge( null, trans2Field3, table2field3, DictionaryConst.LINK_POPULATES );
     graph.addEdge( null, trans2Field4, table2field4, DictionaryConst.LINK_POPULATES );
-
-  }
-
-  public void dumpGraph( Graph g, String fileName ) {
-    // dump the graph to XML for visual inspection etc
-    GraphMLWriter writer = new GraphMLWriter();
-    File file = new File( fileName );
-    OutputStream outputStream = null;
-    try {
-      outputStream = new FileOutputStream( file );
-      writer.outputGraph( g, outputStream );
-    } catch ( Exception e ) {
-    } finally {
-      try {
-        outputStream.close();
-      } catch ( IOException e ) {
-      }
-    }
-
   }
 
   private Vertex createVertex( String id, String type, String name ) {
@@ -144,5 +246,6 @@ public class TestTinkerGraphMetaverseReader extends BaseGraphMetaverseReader {
     v.setProperty( DictionaryConst.PROPERTY_NAME, name );
     return v;
   }
+
 
 }
